requirements.txt
streamlit==1.32.0
langgraph==0.0.26
langchain-core>=0.1.45,<0.2.0
langchain-groq==0.1.3
pydantic==2.6.3
python-dotenv==1.0.1
pypdf==4.1.0
fastapi
uvicorn
